+++ src/portfolio/mod.rs
@@ -5,6 +5,7 @@ mod tracked;
 mod trade;
 mod performance;
 
+use std::collections::HashMap;
 pub use assets::AssetHandlers;
 pub use capital::CapitalHandlers;
 pub use position::PositionHandlers;
@@ -14,7 +15,7 @@ pub use performance::PerformanceMetrics;
 use crate::markets::FeeCalculator;
 use crate::portfolio::tracked::TrackedValue;
 use chrono::{Duration, NaiveDateTime, Utc};
-use polars::prelude::DataFrame;
+use crate::types::{ExecutedTrade, FailedTrade};
 
 pub const DEFAULT_LIMIT: usize = 4;
 pub const DEFAULT_TIMEOUT_MINUTES: i64 = 60 * 2;
@@ -80,8 +81,8 @@ impl Default for PortfolioArgs {
 /// It is responsible for managing the assets and capital available to the portfolio,
 /// as well as the open positions and executed trades.
 pub struct Portfolio {
-    failed_trades: DataFrame,
-    executed_trades: DataFrame,
+    failed_trades: Vec<FailedTrade>,
+    executed_trades: HashMap<NaiveDateTime, ExecutedTrade>,
     open_positions: Vec<NaiveDateTime>,
 
     threshold: f64,
@@ -93,48 +94,56 @@ pub struct Portfolio {
     fee_calculator: Option<Box<dyn FeeCalculator>>,
 }
 
-impl Portfolio {
-    pub fn new<T>(assets: f64, capital: f64, point: T) -> Portfolio
-    where
-        T: Into<Option<NaiveDateTime>>,
-    {
-        let point = point.into().unwrap_or_else(|| {
-            NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap()
-        });
-
-        Portfolio {
-            failed_trades: DataFrame::empty(),
-            executed_trades: DataFrame::empty(),
+impl Default for Portfolio {
+    fn default() -> Self {
+        Self {
+            failed_trades: vec![],
+            executed_trades: HashMap::new(),
             open_positions: vec![],
 
             threshold: DEFAULT_THRESHOLD,
-            assets_ts: TrackedValue::with_initial(assets, point),
-            capital_ts: TrackedValue::with_initial(capital, point),
+            assets_ts: TrackedValue::default(),
+            capital_ts: TrackedValue::default(),
             open_positions_limit: DEFAULT_LIMIT,
             timeout: Duration::minutes(DEFAULT_TIMEOUT_MINUTES),
             fee_calculator: None,
         }
     }
 
+}
+
+impl Portfolio {
+    pub fn new<T>(assets: f64, capital: f64, timestamp: T) -> Portfolio
+    where
+        T: Into<Option<NaiveDateTime>>,
+    {
+        let point = timestamp.into().unwrap_or_else(|| {
+            NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap()
+        });
+
+        Portfolio {
+            assets_ts: TrackedValue::with_initial(assets, point),
+            capital_ts: TrackedValue::with_initial(capital, point),
+            ..Default::default()
+        }
+    }
+
     pub fn from_args(args: &PortfolioArgs, start_time: NaiveDateTime) -> Self {
         Self {
-            failed_trades: DataFrame::empty(),
-            executed_trades: DataFrame::empty(),
-            open_positions: vec![],
-
             threshold: args.threshold,
             assets_ts: TrackedValue::with_initial(args.assets, start_time),
             capital_ts: TrackedValue::with_initial(args.capital, start_time),
             open_positions_limit: args.open_positions_limit,
             timeout: Duration::minutes(args.timeout),
             fee_calculator: None,
+            ..Default::default()
         }
     }
 
     /// Constructor with loaded data
     pub fn with_data(
-        failed_trades: DataFrame,
-        executed_trades: DataFrame,
+        failed_trades: Vec<FailedTrade>,
+        executed_trades: HashMap<NaiveDateTime, ExecutedTrade>,
         open_positions: Vec<NaiveDateTime>,
         assets_ts: TrackedValue,
         capital_ts: TrackedValue,
@@ -235,8 +244,8 @@ mod tests {
         );
 
         // assert that the trade storage is empty
-        assert_eq!(portfolio.executed_trades.height(), 1);
-        assert_eq!(portfolio.failed_trades.height(), 1);
+        assert_eq!(portfolio.executed_trades.len(), 1);
+        assert_eq!(portfolio.failed_trades.len(), 1);
         assert_eq!(portfolio.open_positions.len(), 1);
     }
 
@@ -308,168 +317,4 @@ mod tests {
         assert_eq!(portfolio.timeout, Duration::minutes(10));
     }
 
-    /// Simulates a typical scenario in which a portfolio is created, and then
-    /// a series of trades are executed, some of which are profitable and some of which are not.
-    /// Check to make sure that the portfolio is updated appropriately.
-    #[test]
-    fn market_simulation() {
-        let time = NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap()
-            - Duration::seconds(1);
-
-        let mut portfolio = Portfolio::new(0.0, 300.0, time - Duration::seconds(1));
-        assert_eq!(portfolio.get_assets(), 0.0);
-        assert_eq!(portfolio.available_capital(), 300.0);
-
-        // this will be the sequences of prices used to simulate the market
-        let mut prices = VecDeque::from_iter(&[
-            100.0, // buy
-            99.0,  // buy
-            98.0,  // attempt sell
-            97.0,  // buy
-            98.0,  // sell
-            101.0, // sell
-        ]);
-
-        /*********************
-        handle the first buy
-        *********************/
-
-        // simulate a buy order
-        let price = prices.pop_front().unwrap();
-        let trade = ExecutedTrade::new_without_cost(
-            "id".to_string(),
-            Side::Buy,
-            *price,
-            1.0,
-            time + Duration::milliseconds(1),
-        );
-        portfolio.add_executed_trade(trade);
-
-        // assert that capital and assets have changed accordingly
-        assert_eq!(portfolio.get_assets(), 1.0);
-        assert_eq!(portfolio.available_capital(), 200.0);
-
-        // assert that trade storage, open positions, and available open positions have been updated
-        assert_eq!(portfolio.get_executed_trades().height(), 1);
-        assert_eq!(portfolio.get_open_positions().unwrap().height(), 1);
-        assert_eq!(portfolio.available_open_positions(), DEFAULT_LIMIT - 1);
-
-        /**********************
-        handle the second buy
-        **********************/
-
-        // simulate another buy order at a lower price than the first
-        let price = prices.pop_front().unwrap();
-        let trade = ExecutedTrade::new_without_cost(
-            "id".to_string(),
-            Side::Buy,
-            *price,
-            1.0,
-            time + Duration::milliseconds(2),
-        );
-        portfolio.add_executed_trade(trade);
-
-        // assert that capital and assets have changed accordingly
-        assert_eq!(portfolio.get_assets(), 2.0);
-        assert_eq!(portfolio.available_capital(), 101.0);
-
-        // assert that trade storage, open positions, and available open positions have been updated
-        assert_eq!(portfolio.get_executed_trades().height(), 2);
-        assert_eq!(portfolio.get_open_positions().unwrap().height(), 2);
-        assert_eq!(portfolio.available_open_positions(), DEFAULT_LIMIT - 2);
-
-        /*****************************
-        attempt an unprofitable sell
-        *****************************/
-
-        // attempt to generate a sell order using `is_rate_profitable` at a rate which is not profitable
-        let price = prices.pop_front().unwrap();
-        let potential_trade =
-            FutureTrade::new(Side::Sell, *price, 1.0, time + Duration::milliseconds(3));
-        let result = portfolio.is_rate_profitable(potential_trade.get_price());
-
-        // assert that there is no proposed trade
-        assert!(result.is_none());
-
-        // assert that capital and assets have not changed
-        assert_eq!(portfolio.get_assets(), 2.0);
-        assert_eq!(portfolio.available_capital(), 101.0);
-
-        // assert that trade storage, open positions, and available open positions have not been updated
-        assert_eq!(portfolio.get_executed_trades().height(), 2);
-        assert_eq!(portfolio.get_open_positions().unwrap().height(), 2);
-        assert_eq!(portfolio.available_open_positions(), DEFAULT_LIMIT - 2);
-
-        /*********************
-        handle the third buy
-        *********************/
-
-        // simulate another buy order at a lower price than the second
-        let price = prices.pop_front().unwrap();
-        let trade = ExecutedTrade::new_without_cost(
-            "id".to_string(),
-            Side::Buy,
-            *price,
-            1.0,
-            time + Duration::milliseconds(4),
-        );
-        portfolio.add_executed_trade(trade);
-
-        // assert that capital and assets have changed accordingly
-        assert_eq!(portfolio.get_assets(), 3.0);
-        assert_eq!(portfolio.available_capital(), 4.0);
-
-        // assert that trade storage, open positions, and available open positions have been updated
-        assert_eq!(portfolio.get_executed_trades().height(), 3);
-        assert_eq!(portfolio.get_open_positions().unwrap().height(), 3);
-        assert_eq!(portfolio.available_open_positions(), DEFAULT_LIMIT - 3);
-
-        /**************************
-        attempt a profitable sell
-        **************************/
-
-        // generate a sell order using `is_rate_profitable` at a rate which would sell the third buy order
-        let price = prices.pop_front().unwrap();
-        let potential_trade = portfolio.is_rate_profitable(*price).unwrap();
-
-        assert_eq!(potential_trade.get_side(), Side::Sell);
-        assert_eq!(potential_trade.get_price(), *price);
-        assert_eq!(potential_trade.get_quantity(), 1.0);
-
-        let trade = ExecutedTrade::with_future_trade("id".to_string(), potential_trade);
-        portfolio.add_executed_trade(trade);
-
-        // assert that capital and assets have changed accordingly
-        assert_eq!(portfolio.get_assets(), 2.0);
-        assert_eq!(portfolio.available_capital(), 102.0);
-
-        // assert that trade storage, open positions, and available open positions have been updated
-        assert_eq!(portfolio.get_executed_trades().height(), 4);
-        assert_eq!(portfolio.get_open_positions().unwrap().height(), 2);
-        assert_eq!(portfolio.available_open_positions(), DEFAULT_LIMIT - 2);
-
-        /****************************
-        sell the rest of the assets
-        ****************************/
-
-        // simulate a sell order that will sell the first and second buy
-        let price = prices.pop_front().unwrap();
-        let potential_trade = portfolio.is_rate_profitable(*price).unwrap();
-
-        assert_eq!(potential_trade.get_side(), Side::Sell);
-        assert_eq!(potential_trade.get_price(), *price);
-        assert_eq!(potential_trade.get_quantity(), 2.0);
-
-        let trade = ExecutedTrade::with_future_trade("id".to_string(), potential_trade);
-        portfolio.add_executed_trade(trade);
-
-        // assert that capital and assets have changed accordingly
-        assert_eq!(portfolio.get_assets(), 0.0);
-        assert_eq!(portfolio.available_capital(), 304.0);
-
-        // assert that trade storage, open positions, and available open positions have been updated
-        assert_eq!(portfolio.get_executed_trades().height(), 5);
-        assert!(portfolio.get_open_positions().is_none());
-        assert_eq!(portfolio.available_open_positions(), DEFAULT_LIMIT);
-    }
 }