+++ src/portfolio/trade.rs
@@ -1,3 +1,4 @@
+use std::collections::HashMap;
 use crate::portfolio::assets::AssetHandlers;
 use crate::portfolio::capital::CapitalHandlers;
 use crate::portfolio::position::PositionHandlers;
@@ -9,21 +10,22 @@ use polars::prelude::DataFrame;
 
 /// Interface methods for storing and retrieving trades, and determining when to trade
 pub trait TradeHandlers: PositionHandlers + AssetHandlers + CapitalHandlers {
-    fn get_executed_trades(&self) -> &DataFrame;
+    fn get_executed_trades(&self) -> &HashMap<NaiveDateTime, ExecutedTrade>;
     fn add_failed_trade(&mut self, trade: FailedTrade);
     fn add_executed_trade(&mut self, trade: ExecutedTrade);
-    #[deprecated(note="Use generate_sell_opt() instead")]
-    fn is_rate_profitable(&self, rate: f64) -> Option<FutureTrade>;
+
+    #[deprecated(note="Responsibility is moving to crate::PositionManager")]
     fn generate_sell_opt(&self, candle: &Candle) -> Option<FutureTrade>;
+    #[deprecated(note="Responsibility is moving to crate::PositionManager")]
     fn generate_buy_opt(&self, candle: &Candle) -> Option<FutureTrade>;
     fn get_buy_cost(&self) -> f64;
-    fn get_last_trade(&self) -> Option<ExecutedTrade>;
+    fn get_last_trade(&self) -> Option<&ExecutedTrade>;
     fn able_to_buy(&self) -> bool;
 }
 
 impl TradeHandlers for Portfolio {
     /// Get the executed trades
-    fn get_executed_trades(&self) -> &DataFrame {
+    fn get_executed_trades(&self) -> &HashMap<NaiveDateTime, ExecutedTrade> {
         &self.executed_trades
     }
 
@@ -34,8 +36,7 @@ impl TradeHandlers for Portfolio {
     /// # Arguments
     /// * `trade` - The failed trade to add
     fn add_failed_trade(&mut self, trade: FailedTrade) {
-        let row = trade.as_dataframe();
-        self.failed_trades = self.failed_trades.vstack(&row).unwrap();
+        self.failed_trades.push(trade)
     }
 
     /// Add an executed trade to the portfolio
@@ -52,61 +53,21 @@ impl TradeHandlers for Portfolio {
         } else {
             self.increase_capital(trade.get_cost(), *trade.get_point());
             self.decrease_assets(trade.get_quantity(), *trade.get_point());
-            self.clear_open_positions(&trade)
+            self.clear_open_positions(&trade);
         }
-        let row = trade.as_dataframe();
-        self.executed_trades = self.executed_trades.vstack(&row).unwrap();
-    }
-
-    /// Check if a given rate is profitable to sell assets, then return a `FutureTrade` with the appropriate
-    /// side and quantity.
-    ///
-    /// This method is used to determine if a given rate is profitable to sell assets at. If it is, then
-    /// open positions are selected with the `Portfolio::select_open_positions()` method and the quantity
-    /// of the trade is set to the sum of the quantities of the selected open positions.
-    ///
-    /// # Arguments
-    /// * `rate` - The proposed rate to check
-    ///
-    /// # Returns
-    /// * `Some` - A `FutureTrade` with the appropriate side, quantity, and rate
-    /// * `None` - If the rate is not profitable
-    fn is_rate_profitable(&self, rate: f64) -> Option<FutureTrade> {
-        let viable_positions = self.select_open_positions(rate);
-
-        if let Some(positions) = viable_positions {
-            // the total quantity of assets to be sold
-            let quantity: f64 = positions.column("quantity").unwrap().sum().unwrap();
-
-            // the total cost at which the assets were purchased
-            let cost: f64 = positions.column("cost").unwrap().sum().unwrap();
-
-            // calculate the value of the assets at the proposed rate
-            let sell_value = match self.fee_calculator {
-                Some(ref fee_calculator) => {
-                    fee_calculator.cost_including_fee(quantity * rate, Side::Sell)
-                }
-                None => quantity * rate,
-            };
-            let profit = sell_value - cost;
-            if profit > self.threshold {
-                let time = NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap();
-                return Some(FutureTrade::new(Side::Sell, rate, quantity, time));
-            }
-        }
-        None
+        self.executed_trades.insert(*trade.get_point(), trade);
     }
 
     fn generate_sell_opt(&self, candle: &Candle) -> Option<FutureTrade> {
         let rate = calculate_sell_rate(candle);
-        let viable_positions = self.select_open_positions(rate);
+        let viable_positions = self.select_open_positions_by_price(rate);
 
         if let Some(positions) = viable_positions {
             // the total quantity of assets to be sold
-            let quantity: f64 = positions.column("quantity").unwrap().sum().unwrap();
+            let quantity: f64 = positions.iter().map(|x| x.get_quantity()).sum();
 
             // the total cost at which the assets were purchased
-            let cost: f64 = positions.column("cost").unwrap().sum().unwrap();
+            let cost: f64 = positions.iter().map(|x| x.get_cost()).sum();
 
             // calculate the value of the assets at the proposed rate
             let sell_value = match self.fee_calculator {
@@ -145,11 +106,10 @@ impl TradeHandlers for Portfolio {
     /// # Returns
     /// * `Some` - The most recent trade
     /// * `None` - If there are no trades
-    fn get_last_trade(&self) -> Option<ExecutedTrade> {
-        if self.executed_trades.height() > 0 {
-            let last_row = self.executed_trades.tail(Some(1));
-            let trade = ExecutedTrade::from_row(&last_row);
-            Some(trade)
+    fn get_last_trade(&self) -> Option<&ExecutedTrade> {
+        if self.executed_trades.len() > 0 {
+            let last_ts = self.executed_trades.keys().max().unwrap();
+            self.executed_trades.get(last_ts)
         } else {
             None
         }
@@ -221,7 +181,7 @@ mod tests {
             NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap(),
         );
         portfolio.add_failed_trade(trade);
-        assert_eq!(portfolio.failed_trades.height(), 1);
+        assert_eq!(portfolio.failed_trades.len(), 1);
 
         // add a failed sell
         let trade = FailedTrade::new(
@@ -232,7 +192,7 @@ mod tests {
             NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap(),
         );
         portfolio.add_failed_trade(trade);
-        assert_eq!(portfolio.failed_trades.height(), 2);
+        assert_eq!(portfolio.failed_trades.len(), 2);
     }
 
     /// Test that an executed trade is correctly added to the portfolio storage
@@ -254,7 +214,7 @@ mod tests {
         assert_eq!(portfolio.get_assets(), 200.0);
 
         portfolio.add_executed_trade(trade);
-        assert_eq!(portfolio.executed_trades.height(), 1);
+        assert_eq!(portfolio.executed_trades.len(), 1);
         assert_eq!(portfolio.open_positions.len(), 1);
 
         // check that capital and assets are updated
@@ -267,86 +227,15 @@ mod tests {
             Side::Sell,
             100.0,
             1.0,
-            NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap(),
+            (Utc::now() + Duration::seconds(1)).naive_utc()
         );
-
         portfolio.add_executed_trade(trade);
-        assert_eq!(portfolio.executed_trades.height(), 2);
+        assert_eq!(portfolio.executed_trades.len(), 2);
 
         // check that capital and assets are updated
         assert_eq!(portfolio.available_capital(), 200.0);
         assert_eq!(portfolio.get_assets(), 200.0);
         assert_eq!(portfolio.open_positions.len(), 0);
-    }
-
-    #[test]
-    fn test_is_rate_profitable() {
-        let mut portfolio = Portfolio::new(200.0, 200.0, None);
-        let trade_price = 100.0;
-
-        // check that a rate is not profitable when there are no open positions
-        assert!(portfolio.is_rate_profitable(trade_price).is_none());
-
-        // add an open position
-        let price = 90.0;
-        let quantity = 1.0;
-        let trade = ExecutedTrade::new_without_cost(
-            "id".to_string(),
-            Side::Buy,
-            price,
-            quantity,
-            NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap(),
-        );
-        portfolio.add_executed_trade(trade);
-        assert_eq!(portfolio.open_positions.len(), 1);
-        assert_eq!(portfolio.executed_trades.height(), 1);
-
-        // check that a properly formatted trade is returned when rate is profitable
-        let proposed_trade = portfolio.is_rate_profitable(trade_price).unwrap();
-        assert_eq!(proposed_trade.get_quantity(), quantity);
-        assert_eq!(proposed_trade.get_price(), trade_price);
-
-        // assert that a non-profitable trade is not included in the proposed trade
-        let trade2 = ExecutedTrade::new_without_cost(
-            "id".to_string(),
-            Side::Buy,
-            100.1,
-            quantity,
-            NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap(),
-        );
-        portfolio.add_executed_trade(trade2);
-
-        // check that the non-profitable trade is not included
-        let proposed_trade = portfolio.is_rate_profitable(trade_price).unwrap();
-        assert_eq!(proposed_trade.get_price(), trade_price);
-        assert_eq!(proposed_trade.get_quantity(), quantity);
-
-        let trade3 = ExecutedTrade::new_without_cost(
-            "id".to_string(),
-            Side::Buy,
-            90.0,
-            quantity,
-            NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap(),
-        );
-        portfolio.add_executed_trade(trade3);
-
-        let proposed_trade = portfolio.is_rate_profitable(trade_price).unwrap();
-        assert_eq!(proposed_trade.get_price(), trade_price);
-        assert_eq!(proposed_trade.get_quantity(), quantity * 2.0);
-    }
-
-    /// This test is identical to the one above, except that it uses a fee calculator
-    #[test]
-    fn test_is_rate_profitable_denies_trade() {
-        let time = NaiveDateTime::from_timestamp_opt(Utc::now().timestamp(), 0).unwrap();
-        let mut portfolio = Portfolio::new(200.0, 200.0, time - Duration::seconds(1))
-            .add_fee_calculator(crate::markets::SimplePercentageFee::new(1.0));
-        portfolio.set_threshold(1.0);
-
-        // build a buy trade
-        let trade = ExecutedTrade::new_without_cost("id".to_string(), Side::Buy, 100.0, 1.0, time);
-
-        assert!(portfolio.is_rate_profitable(100.1).is_none());
     }
 
     #[test]